requirements.txt
pandas>=2.0
sympy>=1.12
streamlit>=1.36
python-dotenv>=1.0
openai>=1.43.0
openpyxl>=3.1
